fix(recap): Moved recap email tests to test_recap_email
